Добавил валидацию + рефакторинг

+++ src/main/java/ru/saidgadjiev/aboutme/controller/LikeController.java
@@ -3,6 +3,7 @@ package ru.saidgadjiev.aboutme.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import ru.saidgadjiev.aboutme.model.LikeDetails;
 import ru.saidgadjiev.aboutme.service.LikeService;
 
+import javax.validation.Valid;
 import java.sql.SQLException;
 
 @RestController
@@ -21,7 +23,13 @@ public class LikeController {
 
     @PreAuthorize("isAuthenticated()")
     @RequestMapping(value = "/like/post", method = RequestMethod.POST)
-    public ResponseEntity<LikeDetails> like(@RequestBody LikeDetails likeDetails) throws SQLException {
+    public ResponseEntity<LikeDetails> like(
+            @RequestBody @Valid LikeDetails likeDetails,
+            BindingResult bindingResult
+    ) throws SQLException {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.badRequest().build();
+        }
         likeService.create(likeDetails);
         likeDetails.setLikesCount((int) likeService.postLikes(likeDetails.getPostId()));
         likeDetails.setLiked(true);
@@ -31,7 +39,13 @@ public class LikeController {
 
     @PreAuthorize("isAuthenticated()")
     @RequestMapping(value = "/dislike/post", method = RequestMethod.POST)
-    public ResponseEntity<LikeDetails> dislike(@RequestBody LikeDetails likeDetails) throws SQLException {
+    public ResponseEntity<LikeDetails> dislike(
+            @RequestBody @Valid LikeDetails likeDetails,
+            BindingResult bindingResult
+    ) throws SQLException {
+        if (bindingResult.hasErrors()) {
+            return ResponseEntity.badRequest().build();
+        }
         likeService.remove(likeDetails);
         likeDetails.setLikesCount((int) likeService.postLikes(likeDetails.getPostId()));
         likeDetails.setLiked(false);
